fix problem when adding compiler flags

+++ srcs/utility/ft_is_valid_pos.c
@@ -6,7 +6,7 @@
 /*   By: ktieu <ktieu@student.hive.fi>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/02/13 11:59:32 by ktieu             #+#    #+#             */
-/*   Updated: 2025/02/20 12:54:54 by ktieu            ###   ########.fr       */
+/*   Updated: 2025/02/27 14:03:49 by ktieu            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -21,8 +21,8 @@ int	ft_is_valid_pos(t_cub *c, int32_t x, int32_t y)
 	int32_t	row;
 	int32_t	col;
 
-	if (x < 0 || (uint32_t)x >= c->map.width
-		|| y < 0 || (uint32_t)y >= c->map.height
+	if (x < 0 || x >= c->map.width
+		|| y < 0 || y >= c->map.height
 	)
 		return (0);
 	row = y / CELL_PX;
